doc and tests now maintained on gh-pages branch
GitHub pages is a very convenient way of serving a website for a
project; they can automatically generate HTML from markdown, have syntax
highlighting, etc. and it's all updated automatically when you push to
the gh-pages branch.

I've also put the tests there, but made them reference the latest
version of the PlotKit javascript on the master branch. This will enable
us to start having continuous integration tests, since a push to the
repository is all that is needed to bring everything up-to-date.