Extract common code of tests into layout file
GitHub uses the Liquid templating engine to render files on the gh-pages
branch into a website. I have used this to reduce duplication of headers
and footers between the tests. A convenient side-effect is that there is
now one place (_layouts/test.html) where the base URL for scripts is
specified; this makes it possible to easily override the URL for local
testing while still making it reference the head of the master branch on
GitHub for the public view.