Code cleanup in line chart layouting
I removed the condition 'if (this.style != "area")' because the "area"
chart type does not exist anywhere else. I'm not sure if this was
accidentally left in the code after some refactoring in the past, or if
it was the beginnings of a new feature. In any case, I think that it
makes sense that the points in a dataset should be sorted before
connecting them up with a line, whether or not this creates an area
graph. Unless you want to be able to create graphs which wrap back on
themselves on the x axis... but that should probably be a completely
different layout type anyway.