Better _evaluateLimits behaviour with xAxis option
Previously _evaluateLimits searched for the globally maximum x and y
values in the datasets, and adjusted the boundaries of the axes
accordingly. However, this is does not always do what you would expect:
Imagine, for example, you've got the dataset of a stock price over
several years, and you've set the xAxis option to zoom in on a small
area of that data. The variation of the value over that short period
would be much less than the variation over the course of serveral years,
and accordingly you would expect the scale to be adjusted to the
selected xAxis range.

I have changed the behaviour of _evaluateLimits so that it takes the
xAxis option into account unless a yAxis option is also specified.