Implemented clipping to the yAxis range
If a yAxis range is specified as an option, and the data values went
beyond that range, PlotKit previously bounded the points to the yAxis
range, meaning the gradients of the lines before and after that point
were wrong. I have now implemented a proper clipping algorithm which
adds additional points if necessary, representing a cut-off/flattened
curve.